moved views and functions related to Metric page to separate file
